# API-documentation/Blog Application.postman_collection.json-OpenApi31Json.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Blog Application",
    "contact": {},
    "version": "1.0"
  },
...
  "servers": [
    {
      "url": "http://example.com/api/v1",
      "variables": {}
    }
  ],
  "paths": {
...
    "/articles/6192f3db412d6810b6bfe002": {
      "get": {
        "tags": [
          "Articles"
        ],
        "summary": "Get single Article",
        "operationId": "GetsingleArticle",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        },
        "deprecated": false
      },
      "patch": {
        "tags": [
          "Articles"
        ],
        "summary": "Update Article",
        "operationId": "UpdateArticle",
        "parameters": [],
        "requestBody": {
          "description": "",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateArticleRequest"
              },
              "example": {
                "name": "Test Article I"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        },
        "deprecated": false
      },
      "parameters": []
    },
...
    "/comments": {
      "get": {
        "tags": [
          "Comments"
        ],
        "summary": "Get All Comments",
        "operationId": "GetAllComments",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        },
        "deprecated": false
      },
      "post": {
        "tags": [
          "Comments"
        ],
        "summary": "Create Comment",
        "operationId": "CreateComment",
        "parameters": [],
        "requestBody": {
          "description": "",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCommentRequest"
              },
              "example": {
                "comment": "AWESOME",
                "rating": 5,
                "article": "6192f3db412d6810b6bfe002",
                "user": "6192f7f7d1904ba85e17c3b7"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        },
        "deprecated": false
      },
      "parameters": []
    },
    "/comments/6192f81cd1904ba85e17c3bd": {
      "get": {
        "tags": [
          "Comments"
        ],
        "summary": "Get Single Comment",
        "operationId": "GetSingleComment",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        },
        "deprecated": false
      },
      "patch": {
        "tags": [
          "Comments"
        ],
        "summary": "Update Comment",
        "operationId": "UpdateComment",
        "parameters": [],
        "requestBody": {
          "description": "",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateCommentRequest"
              },
              "example": {
                "rating": 5
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        },
        "deprecated": false
      },
      "delete": {
        "tags": [
          "Comments"
        ],
        "summary": "Delete Comment",
        "operationId": "DeleteComment",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        },
        "deprecated": false
      },
      "parameters": []
    },
...
    "/users/6192f7f7d1904ba85e17c3b7": {
      "get": {
        "tags": [
          "Users"
        ],
        "summary": "Get single User",
        "operationId": "GetsingleUser",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        },
        "deprecated": false
      },
      "patch": {
        "tags": [
          "Users"
        ],
        "summary": "Update User",
        "operationId": "UpdateUser",
        "parameters": [],
        "requestBody": {
          "description": "",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateUserRequest"
              },
              "example": {
                "name": "User Satu"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        },
        "deprecated": false
      },
      "parameters": []
    },
...
  },
  "components": {
    "schemas": {
...
      "UpdateArticleRequest": {
        "title": "UpdateArticleRequest",
        "required": [
          "name"
        ],
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          }
        },
        "examples": [
          {
            "name": "Test Article I"
          }
        ]
      },
      "CreateCommentRequest": {
        "title": "CreateCommentRequest",
        "required": [
          "comment",
          "rating",
          "article",
          "user"
        ],
        "type": "object",
        "properties": {
          "comment": {
            "type": "string"
          },
          "rating": {
            "type": "integer",
            "contentEncoding": "int32"
          },
          "article": {
            "type": "string"
          },
          "user": {
            "type": "string"
          }
        },
        "examples": [
          {
            "comment": "AWESOME",
            "rating": 5,
            "article": "6192f3db412d6810b6bfe002",
            "user": "6192f7f7d1904ba85e17c3b7"
          }
        ]
      },
      "UpdateCommentRequest": {
        "title": "UpdateCommentRequest",
        "required": [
          "rating"
        ],
        "type": "object",
        "properties": {
          "rating": {
            "type": "integer",
            "contentEncoding": "int32"
          }
        },
        "examples": [
          {
            "rating": 5
          }
        ]
      },
...
      "UpdateUserRequest": {
        "title": "UpdateUserRequest",
        "required": [
          "name"
        ],
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          }
        },
        "examples": [
          {
            "name": "User Satu"
          }
        ]
      },
...
    },
    "securitySchemes": {
      "httpBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  },
  "security": [
    {
      "httpBearer": []
    }
  ],
...
}